backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tiktoklive>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
